requirements.txt
click
sqlalchemy
requests
Pillow
